backend/requirements.txt
# Core FastAPI and GraphQL
fastapi==0.104.1
strawberry-graphql[fastapi]==0.214.1
uvicorn[standard]==0.24.0

# ML Frameworks
transformers[torch,tf]==4.35.2
torch==2.1.1
tensorflow==2.15.0
datasets==2.14.6
accelerate==0.24.1

# Quantization
bitsandbytes==0.41.2
onnx==1.15.0
onnxruntime==1.16.3

# File watching for hot reload
watchdog==3.0.0

# Configuration and validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Monitoring
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Utilities
scikit-learn==1.3.2
numpy==1.24.3
